Include explicit information about Java 6
Arduino 1.5.8 requires Java 6 to be installed; on a Fresh Mac OS X installation, this may not be an issue, but if a user has upgraded Java or changed their JAVA_HOME to point to Java 7, the Arduino IDE will crash upon launch.  Including this instruction to save others from the pain I endured.